requirements.txt
backports.shutil-get-terminal-size==1.0.0
behave==1.2.5
decorator==4.0.10
EasyProcess==0.2.3
enum34==1.1.6
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
parse==1.6.6
parse-type==0.3.4
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
ptyprocess==0.5.1
Pygments==2.1.3
PyHamcrest==1.9.0
python-slugify==1.2.1
selenium==3.0.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
Unidecode==0.4.19
wcwidth==0.1.7
xvfbwrapper==0.2.8
